+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -118,7 +118,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KRG5HGSSFEC84W2SWYZWDZDT"),
+  _id: z.string().default("01KRGC9RQPBWCM8WY0S03FCFDM"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,